+++ oci/crud_helpers.go
@@ -527,6 +527,7 @@ func convertResourceFieldsToDatasourceFields(resourceSchema *schema.Resource) *s
 		fieldSchema.Optional = false
 		fieldSchema.DiffSuppressFunc = nil
 		fieldSchema.ValidateFunc = nil
+		fieldSchema.ConflictsWith = nil
 		fieldSchema.Default = nil
 		if fieldSchema.Type == schema.TypeSet {
 			fieldSchema.Type = schema.TypeList