+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Support for the Instance Pools & Instance Configurations
+- Support for the Block Volume cross-region backups
 
 ## 3.8.0 (November 28, 2018)